textbackend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
ccxt==4.4.12
APScheduler==3.10.4
python-dotenv==1.0.0
websockets==12.0
TA-Lib==0.4.24
pytest==7.4.3
